pyproject.toml
[project]
name = "mantelware"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.1.8",
    "pyyaml>=6.0.2",
    "xdg-base-dirs>=6.0.2",
]
